+++ packages/typography/src/styles.tsx
@@ -1,0 +1,13 @@
+import { css } from '@leafygreen-ui/emotion';
+
+export const typeScale1 = css`
+  font-size: 14px;
+  line-height: 20px;
+  letter-spacing: 0px;
+`;
+
+export const typeScale2 = css`
+  font-size: 16px;
+  line-height: 24px;
+  letter-spacing: 0px;
+`;
